End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_modules/paginations';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   loading = false;
 
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,9 +36,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
-    })
+    this.confirmService.confirm('Are you sure you wan to delete?', 'This cannot be undone')
+      .subscribe(result => {
+        if (result){
+          this.messageService.deleteMessage(id).subscribe(()=>{
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+          })
+        }
+      })
+
   }
